+++ kong/plugins/karna/handler.lua
@@ -1,6 +1,6 @@
 local plugin = {
   PRIORITY = 8300,
-  VERSION = "1.4.2",
+  VERSION = "1.4.3",
 }
 
 local ngx                 = ngx